Refactor getter and setters to cache and set when necessary
Because the order of setup seems crucial to the Malibu and II firmware it seems
easier to have cached settings that are all applied on ``init`` or ``acquire``.
This is closer the II examples we are workign off where they pull all the
settings from the GUI and then apply every time.

I also removed some now unused methods such as ``set_defaults``, ``set_routes``,
``set_clocks`` and no longer support setting the sample rate.  We'll just run at
full rate for now.